+++ lib/contextio/folder_collection.rb
@@ -11,8 +11,23 @@ class ContextIO
     belongs_to :source
 
     def create(folder_name, folder_delimiter='/')
-      api.request(:put, "#{resource_url}/#{folder_name}", delim: folder_delimiter)['success']
+      # Original 
+      # => api.request(:put, "#{resource_url}/#{folder_name}", delim: folder_delimiter)['success']
+
+      # Solution? 
+      # => api.request(:put, "#{resource_url}/#{folder_name}?delim=#{folder_delimiter}", delim: folder_delimiter)['success']
+
+      # Hotfix - 
+      # => UNCOMMENTED BELOW
+
+      # => Explanation START:
+      # => Currently the folder delimiter parameter causes an error, took the parameter out and this
+      # => method works now but this method cannot utilize the folder delimiter parameter functionality.
+      # => END
+
+      api.request(:put, "#{resource_url}/#{folder_name}")['success']
     end
+
   end
 end