+++ CHANGES.md
@@ -3,6 +3,9 @@
 ## Unreleased
 
 * Add 'references' to lazy_attributes in message.rb - Dylan Stamat
+* Add remove_folder method to the Message class in message.rb. - Brandon Chow
+* Hotfix the FolderCollection class's create method in folder_collection.rb : the optional delim parameter causes the api call to break. - Brandon Chow
+* Include ContextIO::API::Resource in EmailAddress class to allow for resource url calls. - Brandon Chow
 
 ## 1.7.2